+++ R/use_rprofile.R
@@ -11,7 +11,11 @@ use_rprofile <- function(project = ".") {
           'options("BiocManager.check_repositories" = FALSE, BiocManager.snapshots = "MRAN")',
           'Sys.umask("0002")',
           'if (interactive() & nchar(system.file(package = "prompt")) > 0) prompt::set_prompt(prompt::prompt_git)',
-          'if (interactive() & nchar(system.file(package = "gert")) > 0) library(gert)'
+          'if (interactive() & nchar(system.file(package = "gert")) > 0) library(gert)',
+          'if (interactive() & nchar(system.file(package = "targets")) > 0) {',
+          '  library(targets)',
+          '  tvn <- function(targets_only = TRUE, ...) targets::tar_visnetwork(targets_only = targets_only, ...)',
+          '}'
         )
       )
     } else {